third_party/WebKit/LayoutTests/http/tests/budget/interfaces.html

<!doctype html>
<html>
  <head>
    <title>Budget API: Verifies that the right Budget API interfaces get exposed.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(function() {
        assert_own_property(Navigator.prototype, 'budget');
        assert_own_property(BudgetService.prototype, 'getCost');
        assert_own_property(BudgetService.prototype, 'getBudget');
        assert_own_property(BudgetService.prototype, 'reserve');
        assert_equals(navigator.budget, navigator.budget);
        assert_true('BudgetService' in self);
      }, 'NavigatorBudget should be exposed and have the expected interface in a Document.');
    </script>
  </body>
</html>
